test(daemon): move bgOutputFixture to testhelpers and migrate the last two liveness tests
Finishes the dedup pass rather than stopping at 3.9%:

- HoldsWorkingThenReady and FreshSpawnFromReady still carried the inline
  setup everything else was extracted from, which was the whole residual
  duplication Sonar measured. Both now use the fixture; the ready-start case
  is served by newBGOutputFixtureInState.
- The fixture itself moves to testhelpers_test.go, where the package's other
  shared mocks live. background_process_test.go goes 706 -> 544 lines and is
  back to containing only tests, which is also what CodeScene was flagging
  when its cohesion score dropped.
